connection manager separate test
